Create directory Plugin/WorkspaceImporter and move associated code inside it.
