+++ Plugin/CMakeLists.txt
@@ -87,7 +87,12 @@ if(NOT MINGW AND WXC_APP)
     message("-- libplugin.so is using RPATH set to ${WXC_LIBS_DIR}")
 endif()
 
-file(GLOB SRCS "*.cpp" "LSP/*.cpp" "ThemeImporters/*.cpp" "wxTerminalCtrl/*.cpp")
+file(GLOB SRCS
+    "*.cpp"
+    "LSP/*.cpp"
+    "ThemeImporters/*.cpp"
+    "WorkspaceImporter/*.cpp"
+    "wxTerminalCtrl/*.cpp")
 
 # Define the output
 add_library(plugin SHARED ${SRCS})
