+++ Plugin/WorkspaceImporter/VisualCppImporter.cpp
